backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
httpx==0.27.2
python-dotenv==1.0.1
email-validator
